+++ src/coq_elpi_HOAS.ml
@@ -342,15 +342,16 @@ type options = {
   nonuniform : bool option;
   reversible : bool option;
   keepunivs : bool option;
+  redflags : CClosure.RedFlags.reds option;
 }
 
-let default_options = {
+let default_options () = {
   hoas_holes = Some Verbatim;
   local = None;
   deprecation = None;
   primitive = None;
   failsafe = false;
-  ppwidth = 80;
+  ppwidth = Option.default 80 (Topfmt.get_margin ());
   pp = Normal;
   pplevel = Constrexpr.LevelSome;
   using = None;
@@ -360,6 +361,7 @@ let default_options = {
   nonuniform = None;
   reversible = None;
   keepunivs = None;
+  redflags = None;
 }
 
 type 'a coq_context = {
@@ -983,6 +985,27 @@ let module_inline_default = { module_inline_unspec with
      | state, Elpi.Builtin.Unspec, gls -> state,Declaremods.DefaultInline,gls)
 }
 
+let reduction_flags = let open API.OpaqueData in let open CClosure in declare {
+  name = "coq.redflags";
+  doc = "Set of flags for lazy, cbv, ... reductions";
+  pp = (fun fmt (x : RedFlags.reds) -> Format.fprintf fmt "TODO");
+  compare = Stdlib.compare;
+  hash = Hashtbl.hash;
+  hconsed = false;
+  constants = [
+    "coq.redflags.all",           all          ;
+    "coq.redflags.allnolet",      allnolet     ;
+    "coq.redflags.beta",          beta         ;
+    "coq.redflags.betadeltazeta", betadeltazeta;
+    "coq.redflags.betaiota",      betaiota     ;
+    "coq.redflags.betaiotazeta",  betaiotazeta ;
+    "coq.redflags.betazeta",      betazeta     ;
+    "coq.redflags.delta",         delta        ;
+    "coq.redflags.zeta",          zeta         ;
+    "coq.redflags.nored",         nored        ;
+];
+}
+
 let get_optionc   = E.Constants.declare_global_symbol "get-option"
 
 let get_options ~depth hyps state =
@@ -1074,6 +1097,13 @@ let get_options ~depth hyps state =
       assert (gl = []);
       NonCumulative ud
     in
+  let get_redflags_option () =
+    match API.Data.StrMap.find_opt "coq:redflags" map with
+    | None -> None
+    | Some (t,depth) ->
+      let _, rd, gl = reduction_flags.Elpi.API.Conversion.readback ~depth state t in
+      assert (gl = []);
+      Some rd in
   {
     hoas_holes =
       begin match get_bool_option "HOAS:holes" with
@@ -1094,6 +1124,7 @@ let get_options ~depth hyps state =
     nonuniform = get_bool_option "coq:nonuniform";
     reversible = get_bool_option "coq:reversible";
     keepunivs = get_bool_option "coq:keepunivs";
+    redflags = get_redflags_option ();
   }
 
 let mk_coq_context ~options state =
@@ -1329,7 +1360,7 @@ and under_coq2elpi_ctx ~calldepth state ctx ?(mk_ctx_item=fun decl -> mk_pi_arro
   in
   let state, coq_ctx, hyps =
       let state, coq_ctx, hyps =
-        aux 0 ~depth:calldepth (mk_coq_context ~options:default_options state) [] state (List.rev ctx) in
+        aux 0 ~depth:calldepth (mk_coq_context ~options:(default_options ()) state) [] state (List.rev ctx) in
       state, coq_ctx, hyps in
   let state, t, gls_t = kont coq_ctx hyps ~depth:(calldepth + List.length hyps) state in
   gls := gls_t @ !gls;
@@ -3382,16 +3413,32 @@ let lp2skeleton ~depth coq_ctx constraints state t =
 
 (* {{{  Declarations.module_body -> elpi ********************************** *)
 
-let rec in_elpi_module_item ~depth path state (name, item) = match item with
-  | Declarations.SFBconst _ ->
-      [GlobRef.ConstRef (Constant.make2 path name)]
-  | Declarations.SFBmind { Declarations.mind_packets = [| _ |] } ->
-      [GlobRef.IndRef (MutInd.make2 path name, 0)]
-  | Declarations.SFBmind _ -> nYI "HOAS SFBmind"
-  | Declarations.SFBmodule mb -> in_elpi_module ~depth state mb
-  | Declarations.SFBmodtype _ -> []
+type module_item =
+  | Module of Names.ModPath.t * module_item list
+  | ModuleType of Names.ModPath.t
+  | Gref of Names.GlobRef.t
+  | Functor of Names.ModPath.t * Names.ModPath.t list
+  | FunctorType of Names.ModPath.t * Names.ModPath.t list
 
-and in_elpi_module : 'a. depth:int -> API.Data.state -> 'a Declarations.generic_module_body -> GlobRef.t list =
+let rec in_elpi_module_item ~depth path state (name, item) =
+  let open Declarations in
+  match item with
+  | SFBconst _ ->
+      [Gref (GlobRef.ConstRef (Constant.make2 path name))]
+  | SFBmind { mind_packets } ->
+      CList.init (Array.length mind_packets) (fun i -> Gref (GlobRef.IndRef (MutInd.make2 path name,i)))
+  | SFBmodule ({ mod_mp; mod_type = NoFunctor _ } as b) -> [Module (mod_mp,in_elpi_module ~depth state b) ]
+  | SFBmodule { mod_mp; mod_type = MoreFunctor _ as l } -> [Functor(mod_mp,functor_params l)]
+  | SFBmodtype { mod_mp; mod_type = NoFunctor _ }  -> [ModuleType mod_mp]
+  | SFBmodtype { mod_mp; mod_type = MoreFunctor _ as l }  -> [FunctorType (mod_mp,functor_params l)]
+
+and functor_params x =
+  let open Declarations in
+  match x with
+  | MoreFunctor(_,{ mod_type_alg = Some (NoFunctor (MEident mod_mp)) },rest) -> mod_mp :: functor_params rest
+  | _ -> [] (* XXX non trivial functors, eg P : X with type a = nat, are badly described (no params) *)
+
+and in_elpi_module : 'a. depth:int -> API.Data.state -> 'a Declarations.generic_module_body -> module_item list =
   fun ~depth state { Declarations.
   mod_mp;             (* Names.module_path *)
   mod_expr;           (* Declarations.module_implementation *)
@@ -3410,9 +3457,8 @@ and in_elpi_module : 'a. depth:int -> API.Data.state -> 'a Declarations.generic_
 let rec in_elpi_modty_item (name, item) = match item with
   | Declarations.SFBconst _ ->
       [ Label.to_string name ]
-  | Declarations.SFBmind { Declarations.mind_packets = [| _ |] } ->
+  | Declarations.SFBmind _ ->
       [ Label.to_string name ]
-  | Declarations.SFBmind _ -> nYI "HOAS SFBmind"
   | Declarations.SFBmodule mb -> in_elpi_modty mb
   | Declarations.SFBmodtype _ -> []