+++ src/coq_elpi_HOAS.mli
@@ -46,6 +46,7 @@ type options = {
   nonuniform : bool option;
   reversible : bool option;
   keepunivs : bool option;
+  redflags : CClosure.RedFlags.reds option;
 }
 
 type 'a coq_context = {
@@ -71,7 +72,7 @@ type 'a coq_context = {
 }
 val mk_coq_context : options:options -> State.t -> empty coq_context
 val get_options : depth:int -> Data.hyps -> State.t -> options
-val default_options : options
+val default_options : unit -> options
 val upcast : [> `Options ] coq_context -> full coq_context
 
 val get_current_env_sigma : depth:int ->
@@ -228,8 +229,16 @@ val in_coq_modpath : depth:int -> term -> Names.ModPath.t
 val modpath : Names.ModPath.t Conversion.t
 val modtypath : Names.ModPath.t Conversion.t
 val module_inline_default : Declaremods.inline Conversion.t
+val reduction_flags : CClosure.RedFlags.reds Conversion.t
 
-val in_elpi_module : depth:int -> State.t -> Declarations.module_body -> GlobRef.t list
+type module_item =
+  | Module of Names.ModPath.t * module_item list
+  | ModuleType of Names.ModPath.t
+  | Gref of Names.GlobRef.t
+  | Functor of Names.ModPath.t * Names.ModPath.t list
+  | FunctorType of Names.ModPath.t * Names.ModPath.t list
+
+val in_elpi_module : depth:int -> State.t -> Declarations.module_body -> module_item list
 val in_elpi_module_type : Declarations.module_type_body -> string list
 
 val coercion_status : coercion_status Conversion.t